+++ Source/Core/Math/Matrix.h
@@ -15,7 +15,7 @@ namespace Math {
 
         Matrix3x3(const Vector3<T>& row0, const Vector3<T>& row1, const Vector3<T>& row2);
         // assignment and comparison
-        T* operator[](size_t row_index) const { return m_mat[row_index]; }
+        T* operator[](int row_index) const { return m_mat[row_index]; }
         bool operator==(const Matrix3x3<T>& rhs) const;
         bool operator!=(const Matrix3x3<T>& rhs) const { return !operator==(rhs); }
         // arithmetic operations
@@ -72,9 +72,9 @@ namespace Math {
         static const Matrix4x4<T> IDENTITY;
         static const Matrix4x4<T> ZERO;
 
-        T* operator[](size_t row_index);
+        T* operator[](int row_index);
 
-        const T* operator[](size_t row_index) const;
+        const T* operator[](int row_index) const;
 
         /** Matrix concatenation using '*'.
          */
